Second attempt at reworking how we store pagination data:
Some Mastodon API methods allow arbitrary ranges, and some (like bookmarks and favourites)
require you to use the specific "next" and "previous" response headers when requesting
additional data. This is tricky for our current implementation of NewsFeedViewModel, which
implements a cusor system.

This patch introduces a TimelinePagination object which either holds a cursorId or
a Pagination object depending on the API requirements. This pagination object is
then used by requestRangeForNewerPage() and requestRangeForOlderPage() to respond with
a suitable RequestRange.

The impact of TimelinePagination is limited to the NewsFeedViewModel, to allow for future
opportunities in simplifying the behavious.